@@ -3,14 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.19.0",
+    "axios": "^0.21.1",
     "material-ui-search-bar": "^0.4.2",
-    "react": "^16.10.2",
-    "react-dom": "^16.10.2",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
     "react-google-location": "^1.2.2",
     "react-load-script": "0.0.6",
-    "react-places-autocomplete": "^7.2.1",
-    "react-router-dom": "^5.1.2",
+    "react-places-autocomplete": "^7.3.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.2.0"
   },
   "scripts": {
